Avoid org-block color clashes
base5 clashes with a few other faces in the block. We could use base6, which would make it similar to how codes look in their respective modes. If we need block text to contrast with non-block text (which I think is a reasonable choice), then we could could use base7 instead, which is a little brighter, but is still close to the original.